为空判断

+++ src/main/java/com/qiein/jupiter/web/service/impl/RoleSourceServiceImpl.java
@@ -3,6 +3,7 @@ package com.qiein.jupiter.web.service.impl;
 import com.qiein.jupiter.constant.CommonConstant;
 import com.qiein.jupiter.exception.ExceptionEnum;
 import com.qiein.jupiter.exception.RException;
+import com.qiein.jupiter.util.CollectionUtils;
 import com.qiein.jupiter.util.NumUtil;
 import com.qiein.jupiter.util.StringUtil;
 import com.qiein.jupiter.web.dao.RoleSourceDao;
@@ -25,7 +26,7 @@ public class RoleSourceServiceImpl implements RoleSourceService {
     private RoleSourceDao roleSourceDao;
 
     @Override
-    public void insert(Integer roleId, String sourceId,Integer companyId) {
+    public void insert(Integer roleId, String sourceId, Integer companyId) {
         //参数校验
         if (NumUtil.isInValid(roleId)) {
             throw new RException(ExceptionEnum.ROLE_ID_IS_NULL);
@@ -33,7 +34,9 @@ public class RoleSourceServiceImpl implements RoleSourceService {
 
 
         List<String> sourceIds = Arrays.asList(sourceId.split(CommonConstant.STR_SEPARATOR));
-        roleSourceDao.delete(roleId);
-        roleSourceDao.batchAddRoleSource(roleId,sourceIds,companyId);
+        if (CollectionUtils.isNotEmpty(sourceIds)) {
+            roleSourceDao.delete(roleId);
+            roleSourceDao.batchAddRoleSource(roleId, sourceIds, companyId);
+        }
     }
 }
